requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.3
coverage==7.10.2
cryptography==45.0.6
cx_Oracle==8.3.0
cycler==0.12.1
dash==3.2.0
dash-bootstrap-components==2.0.3
Flask==3.1.1
fonttools==4.59.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.5
narwhals==2.0.1
nest-asyncio==1.6.0
numpy==2.3.2
oracledb==3.3.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
plotly==6.2.0
pluggy==1.6.0
protobuf==6.31.1
pyarrow==21.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.2
pyodbc==5.2.0
pyparsing==3.2.3
pytest==8.4.1
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.4
retrying==1.4.2
rpds-py==0.27.0
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.42
streamlit==1.48.0
structlog==25.4.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
zipp==3.23.0
